+++ web/src/server_events_dispatch.js
@@ -719,6 +719,7 @@ export function dispatch_normal_event(event) {
                 "send_stream_typing_notifications",
                 "send_private_typing_notifications",
                 "send_read_receipts",
+                "web_navigate_to_sent_message",
             ];
 
             const original_home_view = user_settings.web_home_view;